Moved fixes to missileGuidanceCompat
