+++ addons/miscFixes/patchGM/config.cpp
@@ -26,9 +26,7 @@ class CfgPatches {
             "gm_weapons_launchers_rpg7",
             "gm_weapons_launchers_rpg18",
             "gm_weapons_launchers_strela",
-            "gm_weapons_machinepistols_gm_mp2",
-            "gm_weapons_launchers_maljutka",
-            "gm_vehicles_land_wheeled_brdm2_gc_army_brdm2"
+            "gm_weapons_machinepistols_gm_mp2"
         };
         units[] = {
             "land_gm_euro_bridge_02_20_mainroad",
@@ -59,6 +57,5 @@ class CfgPatches {
 #include "CfgAmmo.hpp"
 #include "CfgCloudlets.hpp"
 #include "CfgEventHandlers.hpp"
-#include "CfgMagazines.hpp"
 #include "CfgVehicles.hpp"
 #include "CfgWeapons.hpp"
