+++ addons/missileGuidanceCompat/patchGM/CfgVehicles.hpp
@@ -26,4 +26,35 @@ class CfgVehicles {
             };
         };
     };
+
+    // Updated 9P133 to SACLOS variant
+    class gm_wheeled_base;
+    class gm_wheeled_APC_base: gm_wheeled_base {
+        class Turrets;
+    };
+    class gm_brdm2_base: gm_wheeled_APC_base {
+        class Turrets: Turrets {
+            class MainTurret;
+        };
+    };
+    class gm_brdm2_9p133_base: gm_brdm2_base {
+        class Turrets: Turrets {
+            class MainTurret: MainTurret {
+                magazines[] = {QGVARMAIN(gm_6Rnd_maljutka_heat_9m14p),QGVARMAIN(gm_6Rnd_maljutka_heat_9m14p),QGVARMAIN(gm_6Rnd_maljutka_heat_9m14p)};
+                class Components;
+                class GunClouds;
+                class GunFire;
+                class Hitpoints;
+                class MGunClouds;
+                class OpticsIn;
+                class Reflectors;
+                class TurnIn;
+                class TurnOut;
+                class Turrets;
+                class TurretSpec;
+                class ViewGunner;
+                class ViewOptics;
+            };
+        };
+    };
 };
