+++ addons/missileGuidanceCompat/patchGM/config.cpp
@@ -7,7 +7,9 @@ class CfgPatches {
         requiredVersion = REQUIRED_VERSION;
         requiredAddons[] = {
             "ace_missileguidance",
-            "gm_weapons_launchers_milan"
+            "gm_weapons_launchers_milan",
+            "gm_weapons_launchers_maljutka",
+            "gm_vehicles_land_wheeled_brdm2_gc_army_brdm2"
         };
         skipWhenMissingDependencies = 1;
         author = "Potato";
